+++ src/gravity/multigrid_fft_approximation.F90
@@ -323,8 +323,8 @@ contains
       implicit none
 
       type(cg_level_connected_T), pointer, intent(in) :: curl !< pointer to a level for which we approximate the solution
-      integer,                      intent(in) :: src  !< index of source in cg%q(:)
-      integer,                      intent(in) :: soln !< index of solution in cg%q(:)
+      integer(kind=4),                     intent(in) :: src  !< index of source in cg%q(:)
+      integer(kind=4),                     intent(in) :: soln !< index of solution in cg%q(:)
 
       integer :: nf, n, nsmoo
       type(cg_list_element),  pointer :: cgl
@@ -486,7 +486,7 @@ contains
       implicit none
 
       type(cg_level_connected_T), pointer, intent(in) :: curl !< pointer to a level for which we approximate the solution
-      integer,                      intent(in) :: soln !< index of solution in cg%q(:)
+      integer(kind=4),                     intent(in) :: soln !< index of solution in cg%q(:)
 
       if (grav_bnd == bnd_periodic .and. (nproc == 1 .or. (associated(curl, coarsest%level) .and. single_base) ) ) then
          call curl%reset_boundaries
@@ -568,7 +568,7 @@ contains
       implicit none
 
       type(cg_level_connected_T), pointer, intent(in) :: fine !< level for which approximate the solution
-      integer,                      intent(in) :: soln !< index of solution in cg%q(:) ! \todo change the name
+      integer(kind=4),                     intent(in) :: soln !< index of solution in cg%q(:) ! \todo change the name
 
       integer                       :: i, j, k, d, lh, g, ib, jb, ibh, jbh, l
       type(cg_level_connected_T), pointer           :: coarse
